Return with correct status code when using Terraform

+++ bin/provision.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-set -e
 
 if [[ "$INFRA_ENVIRONMENT" == "local" ]]; then
     $PROJECT_ROOT/bin/benchmark.sh
@@ -9,6 +8,9 @@ elif [[ "$INFRA_ENVIRONMENT" == "aws" ]]; then
     cd $PROJECT_ROOT/build/infrastructure/aws/
 
     terraform init -backend=true -get=true -upgrade
+    if [ $? -ne 0 ]; then
+      exit 1
+    fi
 
     terraform plan \
         -input=false \
@@ -16,23 +18,23 @@ elif [[ "$INFRA_ENVIRONMENT" == "aws" ]]; then
         -refresh=true \
         -var-file="$PROJECT_ROOT/build/infrastructure/config/aws.tfvars" \
         -var-file="$PROJECT_ROOT/build/infrastructure/config/custom.tfvars"
+    if [ $? -ne 0 ]; then
+      exit 1
+    fi
 
     terraform apply \
         -auto-approve \
         -input=false \
-        "$PROJECT_ROOT/build/infrastructure/aws/aws.tfplan" || true
+        "$PROJECT_ROOT/build/infrastructure/aws/aws.tfplan"
 
-    if [[ "$N" == "1" ]]; then
-        arg="-auto-approve"
-    else
-        arg="-auto-approve"
-    fi
+    status_code="$?"
 
     terraform destroy \
-        $arg \
+        -auto-approve \
         -var-file="$PROJECT_ROOT/build/infrastructure/config/aws.tfvars" \
         -var-file="$PROJECT_ROOT/build/infrastructure/config/custom.tfvars"
 
     cd $PROJECT_ROOT
 
+    exit $status_code
 fi
